+++ src/components/HorizontalRule/HorizontalRule.test.tsx
@@ -13,45 +13,33 @@ describe("HorizontalRule Accessibility", () => {
 });
 
 describe("HorizontalRule", () => {
-  it("Renders the UI snapshot correctly", () => {
-    const tree = renderer.create(<HorizontalRule />).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-
-  it("Renders HorizontalRule component", () => {
+  it("renders HorizontalRule component", () => {
     render(<HorizontalRule />);
     expect(screen.getByRole("separator")).toBeInTheDocument();
   });
 
-  it("Renders with proper custom class", () => {
+  it("renders with proper custom class", () => {
     render(<HorizontalRule className="custom-hr" />);
-    expect(screen.getByRole("separator")).toHaveAttribute(
-      "class",
-      "horizontal-rule custom-hr"
-    );
+    const hrClass = screen.getByRole("separator").getAttribute("class");
+    expect(hrClass).toContain("custom-hr");
   });
 
-  it("Renders with correct width and height values", () => {
-    render(
-      <HorizontalRule className="custom-hr" width="720px" height="5rem" />
-    );
-
-    expect(screen.getByRole("separator")).toHaveAttribute(
-      "style",
-      "width: 720px; height: 5rem;"
-    );
-  });
-
-  it("Renders with correct width and height values", () => {
-    render(<HorizontalRule className="custom-hr" width="80%" align="left" />);
-
-    expect(screen.getByRole("separator")).toHaveAttribute(
-      "style",
-      "width: 80%; height: 2px;"
-    );
-    expect(screen.getByRole("separator")).toHaveAttribute(
-      "class",
-      "horizontal-rule horizontal-rule--left custom-hr"
-    );
+  it("renders the UI snapshot correctly", () => {
+    const basic = renderer.create(<HorizontalRule />).toJSON();
+    const updatedHeight = renderer
+      .create(<HorizontalRule height="5rem" />)
+      .toJSON();
+    const updatedWidth = renderer
+      .create(<HorizontalRule width="720px" />)
+      .toJSON();
+    const alignLeft = renderer.create(<HorizontalRule align="left" />).toJSON();
+    const alignRight = renderer
+      .create(<HorizontalRule align="right" />)
+      .toJSON();
+    expect(basic).toMatchSnapshot();
+    expect(updatedHeight).toMatchSnapshot();
+    expect(updatedWidth).toMatchSnapshot();
+    expect(alignLeft).toMatchSnapshot();
+    expect(alignRight).toMatchSnapshot();
   });
 });