+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Composes the DS `Notification` component with Chakra components.
 - Composes the DS `HelperErrorText` component with Chakra.
 - Adds `useNYPLTheme` hook for consuming applications to use NYPL DS values.
+- Composes the DS `HorizontalRule` component with Chakra.
 
 ## 0.25.1 (October 14, 2021)